feat: move transaction into its own module
