+++ src/transaction/poll.rs
@@ -1,0 +1,9 @@
+pub struct ConfirmationPoll {
+    _tx: String,
+}
+
+impl ConfirmationPoll {
+    pub async fn check(tx_id: &String) -> bool {
+        todo!()
+    }
+}
